+++ tests/providers/amazon/aws/sensors/test_redshift.py
@@ -22,7 +22,7 @@ import unittest
 
 import boto3
 
-from airflow.contrib.sensors.aws_redshift_cluster_sensor import AwsRedshiftClusterSensor
+from airflow.providers.amazon.aws.sensors.redshift import AwsRedshiftClusterSensor
 
 try:
     from moto import mock_redshift
@@ -40,7 +40,7 @@ class TestAwsRedshiftClusterSensor(unittest.TestCase):
             MasterUsername='admin',
             MasterUserPassword='mock_password'
         )
-        if len(client.describe_clusters()['Clusters']) == 0:
+        if not client.describe_clusters()['Clusters']:
             raise ValueError('AWS not properly mocked')
 
     @unittest.skipIf(mock_redshift is None, 'mock_redshift package not present')
